requirements.txt
Django
DjangoRestFramework
Django-import-export
django-cors-headers
jsonfield
django-jsonfield
backports-datetime-fromisoformat
npm
